Documentation

All docstrings!

PolynomialGTM.PolynomialGTM — Module

Provides unofficial implementations (in the form of ODESystem and ODEFunction instances) of a polynomial approximation for longitudinal aircraft dynamics. Specifically, these models approximate NASA's Generic Transport Model – a radio-controlled, sub-scale model aircraft which is used flight control research. These publicly available equations were published by Chakraborty et al.

Extended help

License

MIT License

Copyright (c) 2021 Joe Carpinelli

Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

Exports

GTM

GTMFunction

Imports

Base

Core

DocStringExtensions

LinearAlgebra

Memoize

ModelingToolkit

Symbolics

PolynomialGTM.GTM — Method

GTM(; stm, name)

NASA's Generic Transport Model can be approximated (near select flight conditions) as a polynomial model. Here, GTM is a ModelingToolkit.ODESystem which provides publicly available polynomial approximations for GTM longitudinal flight dynamics.

Extended Help

Initial Conditions

The default initial conditions are one trim condition. Two trim conditions for these polynomial-approximated dynamics are shown below.

trim₁ = [[29.6, deg2rad(9), 0.0, deg2rad(9)],   [deg2rad(0.68), 12.7]]
trim₂ = [[25.0, deg2rad(18), 0.0, deg2rad(18)], [deg2rad(-7.2), 59]]

Usage

model = GTM()

References:

Chakraborty et al

Joe Carpinelli

PolynomialGTM.GTMFunction — Method

GTMFunction(; stm, name, kwargs...)

Returns a DifferentialEquations-compatible ODEFunction for GTM dynamics. The stm, and name keyword arguments are passed to GTM. All other keyword arguments are passed directly to ODEFunction.

Extended Help

Usage

Note that this ODEFunction output has several methods, including an in-place method! Function signatures follow ModelingToolkit and DifferentialEquations conventions.

f = GTMFunction()
let u = randn(4), p = randn(2), t = rand()
    f(u,p,t)
end